+++ apps/esketch/frontend/XDHTML/Frame.cpp
@@ -35,20 +35,20 @@ namespace {
 }
 
 void frame::SetLayout(
-	const char *Id,
+	const str::dString &Id,
 	core::rSession &Session )
 {
 	Session.SetElementLayout( Id, XSLAffix_, layout_::Get );
 }
 
 void frame::SetCasting(
-	const char *Id,
+	const str::dString &Id,
 	core::rSession &Session )
 {
 }
 
 void frame::Display(
-	const char *Id,
+	const str::dString &Id,
 	core::rSession &Session )
 {
 	SetLayout( Id, Session );